Fix integer overflow in LoadBalanceHilbertCurve work arrays
With ParticleWorkWeight=100 and ~20M particles, the work accumulation
overflows 32-bit int:

  TotalWork = 1.16e9 cells + 20e6 particles * 100 = 3.16e9 > INT_MAX (2.15e9)

The overflow produces negative or wrap-around values for TotalWork,
WorkPerProcessor, and BlockDivisions, causing grids to be assigned to
invalid processor numbers.  MPI then fails when trying to communicate
with non-existent ranks.

Fix: widen GridWork, ProcessorWork, TotalWork, WorkThisProcessor,
WorkPerProcessor, and WorkLeft from int to long long in both overloads
of LoadBalanceHilbertCurve.

+++ src/enzo/LoadBalanceHilbertCurve.C
@@ -57,13 +57,13 @@ int LoadBalanceHilbertCurve(HierarchyEntry *GridHierarchyPointer[],
 
   /* Initialize */
   
-  int *GridWork = new int[NumberOfGrids];
+  long long *GridWork = new long long[NumberOfGrids];
   int *NewProcessorNumber = new int[NumberOfGrids];
   hilbert_data *HilbertData = new hilbert_data[NumberOfGrids];
   int *BlockDivisions = new int[NumberOfProcessors];
-  int *ProcessorWork = new int[NumberOfProcessors];
+  long long *ProcessorWork = new long long[NumberOfProcessors];
 
-  int TotalWork, WorkThisProcessor, WorkPerProcessor, WorkLeft;
+  long long TotalWork, WorkThisProcessor, WorkPerProcessor, WorkLeft;
   int i, dim, grid_num, Rank, block_num, Dims[MAX_DIMENSION];
   FLOAT GridCenter[MAX_DIMENSION];
   FLOAT LeftEdge[MAX_DIMENSION], RightEdge[MAX_DIMENSION];
@@ -359,12 +359,12 @@ int LoadBalanceHilbertCurve(grid *GridPointers[], int NumberOfGrids,
   /* Initialize */
   
   NewProcessorNumber = new int[NumberOfGrids];
-  int *GridWork = new int[NumberOfGrids];
+  long long *GridWork = new long long[NumberOfGrids];
   hilbert_data *HilbertData = new hilbert_data[NumberOfGrids];
   int *BlockDivisions = new int[NumberOfProcessors];
-  int *ProcessorWork = new int[NumberOfProcessors];
+  long long *ProcessorWork = new long long[NumberOfProcessors];
 
-  int TotalWork, WorkThisProcessor, WorkPerProcessor, WorkLeft;
+  long long TotalWork, WorkThisProcessor, WorkPerProcessor, WorkLeft;
   int i, dim, grid_num, Rank, block_num, Dims[MAX_DIMENSION];
   FLOAT GridCenter[MAX_DIMENSION];
   FLOAT LeftEdge[MAX_DIMENSION], RightEdge[MAX_DIMENSION];
